Feat: include the add and list methods in the main.rb
